+++ lib/davi.dart
@@ -2,7 +2,6 @@ library davi;
 
 export 'src/cell_background.dart';
 export 'src/cell_text_style.dart';
-export 'src/cell_value_mapper.dart';
 export 'src/cell_icon.dart';
 export 'src/column.dart' hide DaviColumnHelper;
 export 'src/column_id.dart';
@@ -18,7 +17,6 @@ export 'src/cell_collision_behavior.dart';
 export 'src/max_span_behavior.dart';
 export 'src/cell_value_stringify.dart';
 export 'src/cell_painter.dart';
-export 'src/cell_bar.dart';
 export 'src/pin_status.dart';
 export 'src/sort_callback_typedef.dart';
 export 'src/row_callback_typedefs.dart';