+++ lib/src/cell_background.dart
@@ -4,4 +4,13 @@ import 'package:flutter/widgets.dart';
 ///
 /// Used by [DaviColumn].
 typedef CellBackgroundBuilder<DATA> = Color? Function(
-    DATA data, int index, bool hovered);
+    BackgroundBuilderParams params);
+
+class BackgroundBuilderParams<DATA> {
+  BackgroundBuilderParams(
+      {required this.data, required this.rowIndex, required this.hovered});
+
+  final DATA data;
+  final int rowIndex;
+  final bool hovered;
+}