+++ lib/src/cell_listenable_builder.dart
@@ -4,4 +4,12 @@ import 'package:flutter/widgets.dart';
 /// The function receives the row's data ([DATA]) and the row index ([rowIndex])
 /// and returns the corresponding [Listenable].
 typedef DaviCellListenableBuilder<DATA> = Listenable? Function(
-    DATA data, int rowIndex);
+    ListenableBuilderParams<DATA> params);
+
+/// Parameters passed to the [DaviCellListenableBuilder] function.
+class ListenableBuilderParams<DATA> {
+  ListenableBuilderParams({required this.data, required this.rowIndex});
+
+  final DATA data;
+  final int rowIndex;
+}