+++ lib/src/cell_semantics_builder.dart
@@ -5,4 +5,17 @@ import 'package:flutter/widgets.dart';
 ///
 /// Used by [DaviColumn].
 typedef DaviCellSemanticsBuilder<DATA> = SemanticsProperties Function(
-    BuildContext context, DATA data, int index, bool hovered);
+    SemanticsBuilderParams params);
+
+class SemanticsBuilderParams<DATA> {
+  SemanticsBuilderParams(
+      {required this.context,
+      required this.data,
+      required this.rowIndex,
+      required this.hovered});
+
+  final BuildContext context;
+  final DATA data;
+  final int rowIndex;
+  final bool hovered;
+}