+++ lib/src/cell_text_style.dart
@@ -4,4 +4,14 @@ import 'package:flutter/widgets.dart';
 ///
 /// Used by [DaviColumn].
 typedef CellTextStyleBuilder<DATA> = TextStyle? Function(
-    DATA data, int index, bool hovered);
+    TextStyleBuilderParams params);
+
+/// Parameters passed to the [CellTextStyleBuilder] function.
+class TextStyleBuilderParams<DATA> {
+  TextStyleBuilderParams(
+      {required this.data, required this.rowIndex, required this.hovered});
+
+  final DATA data;
+  final int rowIndex;
+  final bool hovered;
+}